Fix bug when detecting on-target compiler for tests
The test code, especially the f2py test code, wants to find an on-target
compiler (meson) to help execute the tests. There is an explicit check
to make sure such a compiler exists before running the test. If it
doesn't exist, then the test is skipped.

The check had a bug where it would treat a subprocess.CalledProcessError
as an indication that the compiler is absent. However this is only
thrown when both check==True and the subprocess returns a non-zero
return code. Since check is explicitly set to False, it will never get
thrown.

Instead, catch a FileNotFoundError. That appears to be what is thrown
when the compiler can't be found.